chore(driver): move cy.focused and cy.root into their own files
