+++ packages/driver/src/cy/commands/querying/index.ts
@@ -1,7 +1,11 @@
+import * as Focused from './focused'
 import * as Querying from './querying'
+import * as Root from './root'
 import * as Within from './within'
 
 export {
+  Focused,
   Querying,
+  Root,
   Within,
 }
